+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG for husky
 
+* dev-develop
+    * BUGFIX      #696 prevent propagation of key events in search component
+
 * 0.23.0 (2016-07-28)
     * FEATURE     #697 Created buttons for label component
     * BUGFIX      #687 fixed thumbnail rendering in datagrid