+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * dev-develop
     * ENHANCEMENT #698 Created published icons for itembox
+    * BUGFIX      #701 Prefixed the slide class of the search component
     * BUGFIX      #696 prevent propagation of key events in search component
 
 * 0.23.0 (2016-07-28)