+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * dev-develop
     * ENHANCEMENT #698 Added responsive-width skin in overlay component
+    * ENHANCEMENT #702 Implemented error handling on column-navigation load
     * BUGFIX      #699 Made url comparison in navigation use url parts an not characters
     * ENHANCEMENT #698 Created published icons for itembox
     * BUGFIX      #701 Prefixed the slide class of the search component