+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 * 0.22.0 (2016-07-21)
     * ENHANCEMENT #692 fixed style bug on add button
+    * ENHANCEMENT #689 removed dead thumbnails view
     * BUGFIX      #686 removed "click," event from ckeditor (notice the ",")
     * BUGFIX      #685 navigation refactoring and made navigation history back sensitive
     * ENHANCEMENT #691 Added badge colors